+++ tests/conftest.py
@@ -271,17 +271,3 @@ def json_patients(demo_data_path):
     with open(demo_data_path) as json_data:
         patients = json.load(json_data)
     return patients
-
-
-@pytest.fixture(scope="function")
-def async_response_obj(test_node, json_patients):
-    """Returns the object written to database when sending a match request to
-    an async server"""
-
-    async_response = {
-        "_id": "async_obj_id",
-        "query_id": "test_query_id",
-        "node": {"id": test_node["_id"], "label": test_node["label"]},
-        "query_patient_id": json_patients[0]["id"],
-    }
-    return async_response